refactor: split empire rail, province tile, and scope state parts (#3878)
Extract empire left-rail button column assembly, province overlay
revealed-tile section body, and AppEventHandlerScope widget lifecycle
into library parts without changing public APIs or player-facing behavior.

Refs #3878
