+++ app/lib/core/services/app_event_handler/app_event_handler_scope_state.dart
@@ -1,0 +1,61 @@
+part of 'app_event_handler_scope.dart';
+
+class _AppEventHandlerScopeState extends ConsumerState<AppEventHandlerScope> {
+  AppEventHandler? _handler;
+  var _bound = false;
+  final List<StreamSubscription<dynamic>> _sessionCommandSubs = [];
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    if (_bound) {
+      return;
+    }
+    _bound = true;
+    final bus = ref.read(appEventBusProvider);
+    _handler = AppEventHandler(
+      bus: bus,
+      navigatorKey: appNavigatorKey,
+      dialogBuilders: _dialogBuilders(),
+      onShowSnackBar: _showSnackBar,
+    );
+    _handler!.bind();
+    _sessionCommandSubs.addAll(_sessionCommandListeners(bus));
+    _logEvent.d('AppEventHandler bound; session command listeners attached');
+  }
+
+  void _showSnackBar(ShowSnackBarEvent event) {
+    final ctx = appNavigatorKey.currentContext;
+    if (ctx == null) {
+      return;
+    }
+    final messenger = ScaffoldMessenger.maybeOf(ctx);
+    if (messenger == null) {
+      return;
+    }
+    messenger.showSnackBar(
+      SnackBar(
+        content: Text(event.message),
+        action: event.actionLabel != null && event.action != null
+            ? SnackBarAction(
+                label: event.actionLabel!,
+                onPressed: event.action!,
+              )
+            : null,
+      ),
+    );
+  }
+
+  @override
+  void dispose() {
+    for (final s in _sessionCommandSubs) {
+      s.cancel();
+    }
+    _sessionCommandSubs.clear();
+    _handler?.unbind();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) => widget.child;
+}
